api/app/requirement.txt
fastapi[standard]==0.124.2
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
requests==2.32.5
stripe==14.0.1
uvicorn==0.38.0
sqlalchemy==2.0.45
mollie-api-python==3.9.1
redis==7.1.0
cryptography==46.0.3
